pyproject.toml
[project]
name = "llm-guided-evolution"
version = "0.1.0"
description = "llm-guided-evolution, virtual envrionment that includes packages necessary for local llm integration & model evalution."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deap==1.4.1",
    "flake8>=7.2.0",
    "furo>=2024.8.6",
    "google-genai>=1.12.1",
    "numpy>=2.0.2",
    "pytest>=8.3.5",
    "sphinx>=8.2.3",
    "torch==2.6.0",
    "torchvision",
    "transformers",
    "uvicorn",
    "fastapi",
    "accelerate",
    "numpy",
    "tqdm",
    "hydra-core==1.3.2",
    "omegaconf",
    "pyyaml"
]
